Test two different ways of using futex.
When mode is naive there is no optimization and it takes about 30 seconds
on my desktop computer. If naive is NOT enabled it's about 10x faster at
about 3 seconds. This is "debug" compile mode.

In --release-fast mode it takes 25 seconds! The reason seems to be that
the "stall" loops were optimized away.